refactor FinancialReport
