+++ src/main/app/FinancialReport/FinancialReportUseCaseFactory.java
@@ -2,25 +2,14 @@ package app.FinancialReport;
 
 import data_access.DAOFactory;
 import data_access.account.UserAccountDataAccessInterface;
-import data_access.authentication.UserSignupDataAccessInterface;
-import entity.account.AccountFactory;
 import entity.account.UserAccount;
-import interface_adaptors.FinancialReport.FinancialReportController;
-import interface_adaptors.FinancialReport.FinancialReportPresenter;
-import interface_adaptors.FinancialReport.FinancialReportViewModel;
-import interface_adaptors.HomePageViewModel;
+import interface_adaptors.financial_report.FinancialReportController;
+import interface_adaptors.financial_report.FinancialReportPresenter;
+import interface_adaptors.financial_report.FinancialReportViewModel;
 import interface_adaptors.ViewManagerModel;
-import interface_adaptors.signup.SignupController;
-import interface_adaptors.signup.SignupPresenter;
-import interface_adaptors.signup.SignupViewModel;
-import use_case.FinancialReport.FinancialReportInteractor;
-import use_case.FinancialReport.FinancialReportOutputBoundary;
-import use_case.signup.SharedAccountSignupInteractor;
-import use_case.signup.SignupInteractor;
-import use_case.signup.SignupOutputBoundary;
+import use_case.financial_report.FinancialReportInteractor;
+import use_case.financial_report.FinancialReportOutputBoundary;
 import view.FinancialReport.FinancialReportView;
-import view.home_page.HomePageView;
-import view.signup.SignupView;
 
 import javax.swing.*;
 import java.io.IOException;
